refactor(ownership): relocate tanstack query code to presentation layer
Redesign queryKeys to build each key by spreading its parent (all/byCarId/
byOwnerId/ownerProfiles/ownerProfilesByIds), replacing the previous
ownerProfilesByCarId/ownerProfilesByCarIdPrefix pair that duplicated the
same literal segments independently.
