+++ src/module/car/ownership/presentation/ui/tables/ownerships/actions-dropdown/content/use-content.ts
@@ -4,8 +4,8 @@ import { useRouter } from 'next/navigation';
 import { useRef } from 'react';
 
 import type { OwnershipDto } from '@/car/ownership/application/dto/ownership';
-import { ownershipPromoteMutationOptions } from '@/car/ownership/infrastructure/tanstack/mutation-options/promote';
-import { ownershipRemoveMutationOptions } from '@/car/ownership/infrastructure/tanstack/mutation-options/remove';
+import { ownershipPromoteMutationOptions } from '@/car/ownership/presentation/tanstack/mutation-options/promote';
+import { ownershipRemoveMutationOptions } from '@/car/ownership/presentation/tanstack/mutation-options/remove';
 import { queryKeys as carQueryKeys } from '@/car/presentation/tanstack/query/keys';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
 import type { DialogModalRef } from '@/ui/dialog-modal/dialog-modal';
